+++ packages/ember-application/index.js
@@ -1,2 +1,1 @@
 export { default as EngineInstance } from './lib/system/engine-instance';
-export { getEngineParent, setEngineParent } from './lib/system/engine-parent';
